requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile 
#
appdirs==1.4.3            # via urlextract
beautifulsoup4==4.7.1     # via urltitle
cachetools==3.1.0         # via urltitle
certifi==2018.11.29       # via miniirc
defusedxml==0.5.0         # via pikepdf
humanize==0.5.1           # via urltitle
idna==2.8                 # via urlextract
lxml==4.3.1               # via pikepdf
miniirc==1.1.2
pikepdf==1.1.0            # via urltitle
ruamel.yaml==0.15.89
soupsieve==1.8            # via beautifulsoup4
uritools==2.2.0           # via urlextract
urlextract==0.9
urltitle==0.2.5
